+++ frontend/src/components/About.jsx
@@ -5,7 +5,7 @@ import { styles } from '../styles';
 import { services } from '../constants';
 import { SectionWrapper } from '../hoc';
 import { fadeIn, textVariant } from '../utils/motion';
-import ServiceCard from './ServiceCard';
+import ServiceCard from './ServiceCard/ServiceCard';
 
 const About = () => {
   return (
