+++ frontend/src/components/ServiceCard/ServiceCard.jsx
@@ -1,36 +1,6 @@
 import { React, useState } from 'react';
 import { motion } from 'framer-motion';
-
-const HoverComponent = ({ object }) => {
-  const formattedItems = [];
-
-  object.forEach((item, index) => {
-    if (typeof item === 'string') {
-      formattedItems.push(
-        <div key={index}>
-          <p>{item}</p>
-          <br />
-        </div>,
-      );
-    } else if (typeof item === 'object') {
-      item.forEach((subitem, subindex) => {
-        formattedItems.push(
-          <a
-            href={subitem.link}
-            key={subitem.title}
-            target="_blank"
-            rel="noreferrer"
-          >
-            {subitem.title}
-          </a>,
-        );
-      });
-    }
-    // Add more conditions for other types if needed
-  });
-
-  return <div className="hover-text">{formattedItems}</div>;
-};
+import HoverComponent from './HoverComponent';
 
 const ServiceCard = ({ index, title, icon, hoverText }) => {
   const [isHovered, setIsHovered] = useState(false);
